Centralized text-writing code in output_stream.*
- All serialize-to-text functions that aren't
  explicitly sample-specific moved into one file.
- Renamed some ToString*() functions for
  consistency.
- Update hlsl_resource_types sample to use shared
  code.
